ethereum/src/eth-py/requirements.txt
web3==5.27.0
python-dotenv==0.20.0